Kaniko builder

+++ src/Commands/DeployCommand.php
@@ -9,6 +9,7 @@ use Laravel\VaporCli\Aws\AwsStorageProvider;
 use Laravel\VaporCli\Docker;
 use Laravel\VaporCli\Git;
 use Laravel\VaporCli\Helpers;
+use Laravel\VaporCli\Kaniko;
 use Laravel\VaporCli\Manifest;
 use Laravel\VaporCli\Path;
 use Laravel\VaporCli\ServeAssets;
@@ -33,6 +34,7 @@ class DeployCommand extends Command
             ->addOption('message', null, InputOption::VALUE_OPTIONAL, 'The message for the commit that is being deployed')
             ->addOption('without-waiting', null, InputOption::VALUE_NONE, 'Deploy without waiting for progress')
             ->addOption('fresh-assets', null, InputOption::VALUE_NONE, 'Upload a fresh copy of all assets')
+            ->addOption('keep-build-dir', null, InputOption::VALUE_NONE, 'Do not delete '.Path::vapor().' directory')
             ->setDescription('Deploy an environment');
     }
 
@@ -54,7 +56,9 @@ class DeployCommand extends Command
             $this->vapor->project(Manifest::id())
         ));
 
-        (new Filesystem())->deleteDirectory(Path::vapor());
+        if (!$this->option('keep-build-dir')) {
+            (new Filesystem())->deleteDirectory(Path::vapor());
+        }
 
         $deployment = $this->handleCancellations($this->vapor->deploy(
             $artifact['id'],
@@ -180,6 +184,19 @@ class DeployCommand extends Command
             Helpers::line();
 
             Helpers::step('<comment>Pushing Container Image</comment>');
+
+            if (Manifest::isUsingKaniko($environment)) {
+                Kaniko::publish(
+                    Path::app(),
+                    Manifest::name(),
+                    $environment,
+                    $artifact['container_registry_token'],
+                    $artifact['container_repository'],
+                    $artifact['container_image_tag']
+                );
+
+                return $artifact;
+            }
 
             Docker::publish(
                 Path::app(),
